refactor: rename the file to `mcr_rslt.py`
