+++ src/mcr_analyzer/io/mcr_rslt.py
